backend/requirements.txt
# ─── FastAPI Server ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ─── ML / Inference ─────────────────────────────────────────────
scikit-learn==1.6.1
numpy==1.26.4
pillow==11.1.0

# ─── Database (optional — app works without MongoDB) ────────────
motor==3.7.0
pymongo==4.10.1

# ─── Utilities ──────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==24.1.0

# ─── For training only (not needed to run the API server) ───────
# tensorflow==2.18.0
# opencv-python-headless==4.10.0.84
# matplotlib==3.9.0
# seaborn==0.13.0
